requirements.txt
django
djangorestframework
markdown
django-filter
pillow
python-dotenv